el bug del exponente de vueltas y el de horas solucioando

+++ src/main/java/main/Control.java
@@ -6,7 +6,7 @@ public class Control {
     private ViewInterface view;
     private ModelCrono cronometro;
 
-    private int vueltas;
+    private double vueltas;
 
     public Control() {
         vueltas = 0;
